requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
blinker==1.7.0
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysqlclient==2.2.1
pillow==10.2.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
